refactor: extract static breadth coordination
